fix country selector
